added config.php for wordpress, but still only err502 or missing file

+++ srcs/requirements/wordpress/conf/config-script.sh
@@ -1,26 +1,44 @@
 #!/bin/bash
 
-wp config create	--allow-root \
-					--dbname="$SQL_DATABASE" \
-					--dbuser="$SQL_USER" \
-					--dbpass="$SQL_PASSWORD" \
-					--dbhost=mariadb \
-					--path="$WP_PATH"
+sed -i "s|listen = /run/php/php7.4-fpm.sock|listen = 0.0.0.0:9000|" /etc/php/7.4/fpm/pool.d/www.conf
 
-wp core install 	--allow-root \
-					--path="$WP_PATH" \ 
-					--title="$WP_TITLE" \
-					--url="$DOMAIN_NAME" \
-					--admin_user="$ADMIN_USER" \
-					--admin_password="$ADMIN_PASSWORD" \
-					--admin_email="$ADMIN_MAIL" \
-		
-wp user create  	--allow-root \
-					--path="$WP_PATH" \
-					"$USER_NAME" \
-					"$USER_EMAIL" \
-					--user_pass="$USER_PWD" \
-					--role=contributor \
+	# Mettre à jour les informations de connexion à la base de données dans wp-config.php
+	if [ -f /var/www/html/wp-config.php ]; then
+		sed -i "s|__MYSQL_DATABASE__|$DATABASE_NAME|g" /var/www/html/wp-config.php
+		sed -i "s|__MYSQL_USER__|$MYSQL_USER|g" /var/www/html/wp-config.php
+		sed -i "s|__MYSQL_PASSWORD__|$MYSQL_PASSWORD|g" /var/www/html/wp-config.php
+		sed -i "s|__DB_HOST__|$DB_HOST|g" /var/www/html/wp-config.php
+	fi
+
+# Vérifier si WordPress n'est pas déjà configuré
+if [ -f /var/www/html/.htaccess ] || [ -d /var/www/html/wp-content/uploads ]; then
+    echo "WordPress is already configured. Skipping setup."
+else
+	#Telecharger WordPress
+	wp core download 	--allow-root \
+						--path=/var/www/html
+	echo "Waiting for MariaDB. . ."
+	until mysqladmin ping -hmariadb -u${MYSQL_USER} -p${MYSQL_PASSWORD} --silent; do
+		sleep 2
+	done
+
+	#Installer WordPress
+	wp core install 	--allow-root \
+						--path=/var/www/html \
+						--title="${WP_TITLE}" \
+						--url=${DOMAIN_NAME} \
+						--admin_user=${WP_ADMIN} \
+						--admin_password=${WP_ADMIN_PASSWORD} \
+						--admin_email="${WP_ADMIN}@student.42.fr"
+
+	#Creer l'utilisateur supplémentaire
+	wp user create  	${WP_USER} \
+						"${WP_USER}@user.com" \
+						--allow-root \
+						--role=author \
+						--user_pass=${WP_USER_PWD} \
+						--path=/var/www/html
+fi
 
 exec php-fpm7.4 -F
